moved routes out of layout
