+++ client/src/Components/Common/Layout/index.js
@@ -1,20 +1,12 @@
 import React from 'react';
-import { Router } from '@reach/router';
-
 import Header from '../../Theme/header';
 import Footer from '../../Theme/footer';
-import {IndexPage, Locations, Drives, About, Test} from '../../../Pages';
 
-const Layout = () => (
+
+const Layout = ({children}) => (
     <div>
       <Header />
-      <Router>
-        <IndexPage path="/" />
-        <Drives path="/drives" />
-        <Locations path="/locations" />
-        <About path="/about" />
-        <Test path="/test" />
-      </Router>
+      {children}
       <Footer />
     </div>
 );
