+++ client/src/Components/Pages/IndexPage.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import Button from '../Common/Button/button'
+
+
+const IndexPage = () => (
+	<div>
+	<h1>Home Page Content</h1>
+	<Button>this is a styled button</Button>
+	<h1>try /locations /about /drives === What is this line for?</h1>
+</div>
+)
+
+export default IndexPage
