--- client/src/Components/Theme/home.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import Button from '../Common/Button/button.js'
-
-const Home = () => (
-    <div>
-		<h1>Home Page Content</h1>
-		<Button>this is a styled button</Button>
-		<h1>try /locations /about /drives === What is this line for?</h1>
-	</div>
-)
-
-export default Home
