--- client/src/Pages/IndexPage.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-import Home from '../Components/Theme/home'
-
-const IndexPage = () => (
-		<Home/>
-)
-
-export default IndexPage
